+++ CometSearch/CometFragmentIndex.h
@@ -41,7 +41,7 @@ private:
    static void GenerateFragmentIndex(ThreadPool *tp);
    static void AddFragments(vector<PlainPeptideIndexStruct>& vRawPeptides,
                             int iWhichThread,
-                            int iWhichPeptide,
+                            size_t iWhichPeptide,
                             int modNumIdx,
                             short siNtermMod,
                             short siCtermMod,
